Fix another dependency on an Impl
Move ReferralError into the ForeverRepository instead
