+++ app/data/data-forever/src/main/kotlin/com/hedvig/android/data/forever/ForeverRepository.kt
@@ -7,6 +7,27 @@ import giraffe.ReferralsQuery
 
 interface ForeverRepository {
   suspend fun getReferralsData(): Either<ErrorMessage, ReferralsQuery.Data>
-  suspend fun updateCode(newCode: String): Either<ForeverRepositoryImpl.ReferralError, String>
+  suspend fun updateCode(newCode: String): Either<ForeverRepository.ReferralError, String>
   suspend fun redeemReferralCode(campaignCode: CampaignCode): Either<ErrorMessage, RedeemReferralCodeMutation.Data?>
+
+  sealed interface ReferralError {
+    data class GeneralError(
+      val message: String?,
+    ) : ReferralError
+
+    data class CodeTooLong(
+      val maxCharacters: Int,
+    ) : ReferralError
+
+    data class CodeTooShort(
+      val minCharacters: Int,
+    ) : ReferralError
+
+    data class MaxUpdates(
+      val maxUpdates: Int,
+    ) : ReferralError
+
+    data object CodeIsEmpty : ReferralError
+    data object CodeExists : ReferralError
+  }
 }
